requirements.txt
# Unsloth and Fine-Tuning
unsloth[windows] @ git+https://github.com/unslothai/unsloth.git
torch
xformers
trl
peft
accelerate
bitsandbytes

# RAG Components
langchain
langchain-huggingface
langchain-chroma
sentence-transformers
chromadb

# Application and Data
streamlit
pandas
tqdm

# Evaluation
evaluate
rouge-score